refactor(ruby): split bundler parsing and provenance seams
